moved CreateDogProfile compoment into Navigation compoment
