+++ components/NavigationBar/NavigationBar.styles.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const IconsContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+`;
